pyproject.toml
[tool.poetry]
name = "bbm"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
monai = {extras = ["nibabel"], version = "^1.1.0"}
pynrrd = "^1.0.0"
pyyaml = "^6.0"
tqdm = "^4.65.0"
jupyter = "^1.0.0"
ipykernel = "^6.21.3"
pre-commit = "^3.2.0"
einops = "^0.6.0"
matplotlib = "^3.7.1"
wandb = "^0.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
